Allow search to start from alternative starting point
The original starting population size is given by Watterson's
equation. If that fails, the algorithm tries again starting from half
that, and repeats if that should fail, too. Halving continues for up
to a factor of 10^4, see the function newton in newton.c.

In addition, this version disallows negative population sizes.