Simplifying & Bug Fixing
This is a drastic reorganization of the code. I deleted code for
- linear algebra
- cross-validation
- bootstrapping
- debugging
and concentrated on latest multiroot-finding method. An important bug
fixed along the way was in popSizes.getNextLevel, where I only accepted
improvements if they were larger than 2, instead of looking for the best
level among all levels and then asking whether it improves the log
likelihood from the previous round by more than 2.

In addition, I cleaned up the interface and overall tried to simplify
the program as much as possible.