Added -w Option
With the -w option the search for optimal population sizes always stars
with Watterson's estimator. This used to be the default behavior of
epos. The new default is that the population sizes found so far are used
as starting point for the new search. The new category added gets as
starting size the size of its left-hand side category, as implemented in
findIniN in popSizes.c. The very first population size is still set to
Watterson's estimator.