Greedy Search Running
This commit includes valgrind-tested versions of greedy.c, exhaustive.c,
and newton.c. The tests are called through the function test in
test.c. The search, which builds upon these functions, is specified in
search.c. The greedy search is running as expected, the exhaustive
search not yet. To run the greedy search, try

./epos -l 10000000 -E 0 data/testNewtonF.dat

However, the results differ compared to the old version.