Likelihood of Negative Population Sizes
If a set of population sizes contains a negative value, the likelihood
of this set is set to -DBL_MAX. However, when printed as a float, this
number is huge and looks ugly. So now a zero is printed instead. Note
also that negative population sizes can only occur when preset levels
are used (-L). During a standard search for the optimal combination of
levels, negative population values are avoided precisely because they
are assigned a negative likelihood.