Clarify Message From Abandoned Searches
Previously it said "no improvement"; this can lead to misunderstanding,
as there the likelihood must improve with every new level added to the
calculation. The explanation for the "no improvement" message is, epos
scores sets of population sizes containing one or more negative values
with a negative likelihood. So, while these sets may have a higher
likelihood than the previous maximum, their true likelihood never enters
the search. To make this more explicit, the error message now reads

found no sets of only positive population sizes with improved log-likelihood

This is a bit verbose, but there we are for now.