Added cross-validation
The input data is split into n data sets. From these, n pairs of
training and test data sets are formed by using one of the n data sets
as test data and merging the remaining n-1 data sets into a training
data set. Given a configuration of break points, compute for each of
these pairs the population sizes on the training set and the likelihood
of these population sizes given the test set. Do this for all n data
sets and average the likelihood. This is returned as the likelihood of
the given configuration of breakpoints. By default, n=5, which can be
changed using the -x option. The logic of this is implemented in xval.c
and in the function compPopSizes in popSizes.c.