Error in Searching
Sometimes the log-likelihood from the cross-validation would improve,
but the constellation gave a nonsense result when applied to the
undivided data set. For example:

 ./epos -s 13 -U -l 10000 ../data/t758.sfs
11	2.30e-01	1.00e+00
3	5.61e+14	3.43e+14
2	5.61e+14	1.00e+00

This is now fixed by checking at every level that

a) the average likelihood from the cross-validation steps increases by
at least the amount set by -c (2 by default)

b) the configuration gives a sensible likelihood when applied to the
full data set

Previously only a) was checked.