Adjust d^2
So far, computation of d^2 included the monomorphic sites. These results
could not be compared to those shown in Table 1 of Lapierre et
al. (2017, Genetics, 206:439), which were computed with respect to the
polymorphic sites. Now my d^2 is also only computed from the polymorphic
sites making the d^2 results comparable.