Fixed error and improved speed without cross-val.
There was an error in the construction of the SFS if no cross validation
was to be carried out (-x 1). Also, with large data sets (e. g. the
Yoruba spectrum) epos takes considerable time to split the spectrum in
preparation for cross-validation. If cross-validation is switched off,
this splitting step is unnecessary, but was nevertheless carried
out. Not any more.