Speed up splitting of site frequency spectra
Site frequency spectra need to be split for cross-validation. So far
this splitting entailed shuffling an array of entries the length of the
input sequence. Even though this splitting needs is carried out only
once, it slows the program drastically down for sequences the size of
the human genome. To speed things up, shuffle only the polymorphic
sites. To compute the number of monomorphic sites, g0, let l be the
length of the input sequence, x the number of cross-validation
categories, and p the number of polymorphic sites in the spectrum. Then
g0 = l / x - p. Apart from time, this also saves memory.